Update plugins, dependencies, and clean up small code inspection issues

+++ core/src/main/java/com/google/zxing/datamatrix/encoder/DefaultPlacement.java
@@ -60,7 +60,7 @@ public class DefaultPlacement {
   }
 
   final void setBit(int col, int row, boolean bit) {
-    bits[row * numcols + col] = bit ? (byte) 1 : (byte) 0;
+    bits[row * numcols + col] = (byte) (bit ? 1 : 0);
   }
 
   final boolean hasBit(int col, int row) {
